The GilaElectromagnetics Module

GilaElectromagnetics — Module

GilaElectromagnetics implements single (complex) frequency electromagnetic Green functions between generalized source and target cuboid ``volumes''. Technical details are available in the supporting document files.

Author: Sean Molesky Distribution: The code distributed under GNU LGPL.

Module Index

GilaElectromagnetics.GlaKerOpt

GilaElectromagnetics.GlaOpr

GilaElectromagnetics.GlaOpr

GilaElectromagnetics.GlaOpr

GilaElectromagnetics.GlaOprMem

GilaElectromagnetics.GlaOprMem

GilaElectromagnetics.GlaVol

GilaElectromagnetics.egoBrnDev!

GilaElectromagnetics.egoBrnHst!

GilaElectromagnetics.genEgoExt!

GilaElectromagnetics.genEgoMat

GilaElectromagnetics.genEgoSlf!

GilaElectromagnetics.glaSze

GilaElectromagnetics.glaSze

GilaElectromagnetics.isadjoint

GilaElectromagnetics.isexternaloperator

GilaElectromagnetics.isselfoperator

Detailed API

GilaElectromagnetics.GlaKerOpt — Method

GlaKerOpt(devStt::Bool)

Simplified GlaKerOpt constructor.

GilaElectromagnetics.GlaOpr — Type

GlaOpr(cel::NTuple{3, Int}, scl::NTuple{3, Rational}, 
org::NTuple{3, Rational}=(0//1, 0//1, 0//1); 
useGpu::Bool=false, setTyp::DataType=ComplexF64)

Construct a self Green operator.

Arguments

cel::NTuple{3, Int}: The number of cells in each dimension.

scl::NTuple{3, Rational}: The size of each cell in each dimension

(in units of wavelength).

org::NTuple{3, Rational}=(0//1, 0//1, 0//1): The origin of the volume in

each dimension (in units of wavelength).

useGpu::Bool=false: Whether to use the GPU (true) or CPU (false).

setTyp::DataType=ComplexF64: The element type of the operator. Must be a

subtype of Complex.

GilaElectromagnetics.GlaOpr — Type

GlaOpr

Abstraction wrapper for GlaOprMem.

Fields

mem::GlaOprMem: Data to process the Green function.

GilaElectromagnetics.GlaOpr — Method

GlaOpr(celSrc::NTuple{3, Int}, sclSrc::NTuple{3, Rational}, 
orgSrc::NTuple{3, Rational}, celTrg::NTuple{3, Int}, 
sclTrg::NTuple{3, Rational}, orgTrg::NTuple{3, Rational}; 
useGpu::Bool=false, setTyp::DataType=ComplexF64)

Construct an external Green's operator.

Arguments

celSrc::NTuple{3, Int}: The number of cells in each dimension of the source

volume.

sclSrc::NTuple{3, Rational}: The size of each cell in each dimension of the

source volume (in units of wavelength).

orgSrc::NTuple{3, Rational}: The origin of the source volume in each

dimension (in units of wavelength).

celTrg::NTuple{3, Int}: The number of cells in each dimension of the target

volume.

sclTrg::NTuple{3, Rational}: The size of each cell in each dimension of the

target volume (in units of wavelength).

orgTrg::NTuple{3, Rational}: The origin of the target volume in each

dimension (in units of wavelength).

useGpu::Bool=false: Whether to use the GPU (true) or CPU (false).

setTyp::DataType=ComplexF64: The element type of the operator. Must be a

subtype of Complex.

GilaElectromagnetics.GlaOprMem — Type

GlaOprMem

Storage structure for a Green function operator. .cmpInf–-computation information see GlaKerOpt .trgVol–-target volume of Green function .srcVol–-source volume of Green function .mixInf–-information for matching source and target grids, see GlaExtInf .dimInfC–-dimension information for Green function volumes, host side .dimInfD–-dimension information for Green function volumes, device side .egoFur–-unique Fourier transform data for circulant Green function .fftPlnFwd–-forward Fourier transform plans .fftPlnRev–-reverse Fourier transform plans .phzInf–-phase vector for splitting Fourier transforms

GilaElectromagnetics.GlaOprMem — Method

function GlaOprMem(cmpInf::GlaKerOpt, trgVol::GlaVol,
srcVol::Union{GlaVol,Nothing}=nothing, 
egoFur::Union{AbstractArray{<:AbstractArray{T}},
Nothing}=nothing)::GlaOprMem where T<:Union{ComplexF64,ComplexF32}

Prepare memory for green function operator–-when called with a single GlaVol, or identical source and target volumes, yields the self construction.

GilaElectromagnetics.GlaVol — Type

GlaVol(cel::Array{<:Integer,1}, celScl::NTuple{3,Rational}, 
org::NTuple{3,Rational}, grdScl::NTuple{3,Rational}=celScl)::GlaVol

Constructor for Gila Volumes.

GilaElectromagnetics.egoBrnDev! — Method

egoBrnDev!(egoMem::GlaOprMem, lvl::Integer, bId::Integer, 
actVec::AbstractArray{T})::AbstractArray{T} where 
T<:Union{ComplexF64,ComplexF32}

Head branching function implementing Green function action on device.

GilaElectromagnetics.egoBrnHst! — Method

egoBrnHst!(egoMem::GlaOprMem, lvl::Integer, bId::Integer, 
actVec::AbstractArray{T})::AbstractArray{T} where 
T<:Union{ComplexF64,ComplexF32}

Head branching function implementing Green function action on host.

GilaElectromagnetics.genEgoExt! — Method

genEgoExt!(egoCrcExt::AbstractArray{T,5}, trgVol::GlaVol, 
srcVol::GlaVol, cmpInf::GlaKerOpt)::Nothing where 
T<:Union{ComplexF64,ComplexF32}

Calculate circulant vector for the Green function between a target volume, trgVol, and source volume, srcVol.

GilaElectromagnetics.genEgoMat — Method

function genEgoMat(celScl::{3,<:Rational}, 
celNum::Ntuple{3,<:Integer})::Array{ComplexF64,2}

Generate dense matrix of a Green function.

GilaElectromagnetics.genEgoSlf! — Method

genEgoSlf!(egoCrc::Array{ComplexF64}, slfVol::GlaVol, 
cmpInf::GlaKerOpt)::Nothing

Calculate circulant vector of the Green function on a single domain.

GilaElectromagnetics.glaSze — Method

glaSze(op::GlaOpr, dim::Int)

Returns the size of the input/output arrays for a GlaOpr in tensor form.

Arguments

op::GlaOpr: The operator to check.

dim::Int: The length of the dimension to check.

Returns

The size of the input/output arrays for a GlaOpr in tensor form.

GilaElectromagnetics.glaSze — Method

glaSze(opr::GlaOpr)

Returns the size of the input/output arrays for a GlaOpr in tensor form.

Arguments

op::GlaOpr: The operator to check.

Returns

A tuple of the sizes of the input and output arrays in tensor form.

GilaElectromagnetics.isadjoint — Method

isadjoint(opr::GlaOpr)

Returns true if the operator is the adjoint of the Greens operator.

Arguments

opr::GlaOpr: The operator to check.

Returns

true if the operator is the adjoint, false otherwise.

GilaElectromagnetics.isexternaloperator — Method

isexternaloperator(opr::GlaOpr)

Returns true if the operator is an external Greens operator.

Arguments

opr::GlaOpr: The operator to check.

Returns

true if the operator is an external Greens operator, false otherwise.

GilaElectromagnetics.isselfoperator — Method

isselfoperator(opr::GlaOpr)

Returns true if the operator is a self Greens operator.

Arguments

opr::GlaOpr: The operator to check.

Returns

true if the operator is a self Greens operator, false otherwise.