The GilaElectromagnetics Module

GilaElectromagnetics — Module

GilaElectromagnetics implements single (complex) frequency electromagnetic Green functions between generalized source and target cuboid ``volumes''. Technical details are available in the supporting document files.

Author: Sean Molesky Distribution: The code distributed under GNU LGPL.

Module Index

GilaElectromagnetics.GlaKerOpt

GilaElectromagnetics.GlaOprMem

GilaElectromagnetics.GlaOprMem

GilaElectromagnetics.GlaVol

GilaElectromagnetics.egoBrnDev!

GilaElectromagnetics.egoBrnHst!

GilaElectromagnetics.genEgoExt!

GilaElectromagnetics.genEgoSlf!

Detailed API

GilaElectromagnetics.GlaKerOpt — Method

GlaKerOpt(devStt::Bool)

Simplified GlaKerOpt constructor.

GilaElectromagnetics.GlaOprMem — Type

GlaOprMem

Storage structure for a Green function operator. .cmpInf–-computation information see GlaKerOpt .trgVol–-target volume of Green function .srcVol–-source volume of Green function .mixInf–-information for matching source and target grids, see GlaExtInf .dimInfC–-dimension information for Green function volumes, host side .dimInfD–-dimension information for Green function volumes, device side .egoFur–-unique Fourier transform data for circulant Green function .fftPlnFwd–-forward Fourier transform plans .fftPlnRev–-reverse Fourier transform plans .phzInf–-phase vector for splitting Fourier transforms

GilaElectromagnetics.GlaOprMem — Method

function GlaOprMem(cmpInf::GlaKerOpt, trgVol::GlaVol,
srcVol::Union{GlaVol,Nothing}=nothing, 
egoFur::Union{AbstractArray{<:AbstractArray{T}},
Nothing}=nothing)::GlaOprMem where T<:Union{ComplexF64,ComplexF32}

Prepare memory for green function operator–-when called with a single GlaVol, or identical source and target volumes, yields the self construction.

GilaElectromagnetics.GlaVol — Type

GlaVol(cel::Array{<:Integer,1}, celScl::NTuple{3,Rational}, 
org::NTuple{3,Rational}, grdScl::NTuple{3,Rational}=celScl)::GlaVol

Constructor for Gila Volumes.

GilaElectromagnetics.egoBrnDev! — Method

egoBrnDev!(egoMem::GlaOprMem, lvl::Integer, bId::Integer, 
actVec::AbstractArray{T})::AbstractArray{T} where 
T<:Union{ComplexF64,ComplexF32}

Head branching function implementing Green function action on device.

GilaElectromagnetics.egoBrnHst! — Method

egoBrnHst!(egoMem::GlaOprMem, lvl::Integer, bId::Integer, 
actVec::AbstractArray{T})::AbstractArray{T} where 
T<:Union{ComplexF64,ComplexF32}

Head branching function implementing Green function action on host.

GilaElectromagnetics.genEgoExt! — Method

genEgoExt!(egoCrcExt::AbstractArray{T,5}, trgVol::GlaVol, 
srcVol::GlaVol, cmpInf::GlaKerOpt)::Nothing where 
T<:Union{ComplexF64,ComplexF32}

Calculate circulant vector for the Green function between a target volume, trgVol, and source volume, srcVol.

GilaElectromagnetics.genEgoSlf! — Method

genEgoSlf!(egoCrc::Array{ComplexF64}, slfVol::GlaVol, 
cmpInf::GlaKerOpt)::Nothing

Calculate circulant vector of the Green function on a single domain.